Add a the ability to build the operator for FIPS-only clusters
This commit includes a file that allows strict usage of algorithms for
FIPS.

This was copied for openshift/boilerplate, which includes some scripts
for enabling FIPS-compliant container image builds. Since we're not
using boilerplate, this just copies the file over, but
ultimately does the same thing as what boilerplate is doing.

To build with strict FIPS runtimes, enable CGO and use the
`strictfipsruntime` Go experimenal variable with a compiler that
supports it.