Fix broken tests as a result of updating controller-runtime 0.16.2
The fake client we use in our unit tests was recently updated to be more
inline with what the actual kubernetes API does:

  https://github.com/kubernetes-sigs/controller-runtime/pull/2259

As a result, some of our tests broken because we were lumping Status and
other updates into the same Update() request.

This commit refactors those tests so that Status updates are implemented
separately from the rest of the runtime objects. We need to do this
before we can upgrade controller-runtime to 0.16.3.